# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a programming assignment for GWI Go Backend position",
        "title": "GWI Go Challenge API",
        "contact": {
            "name": "Dionysis Chasakis",
            "email": "chasakisd@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/audience": {
            "get": {
                "description": "Returns all the application audiences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audience"
                ],
                "summary": "Get audiences",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AudiencePageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates an audience",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audience"
                ],
                "summary": "Create audience",
                "parameters": [
                    {
                        "description": "audienceId",
                        "name": "command",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/response.AudienceCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/audience/favorite": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user's favorite audiences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audience"
                ],
                "summary": "Get favorite audiences",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AudiencePageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/audience/{audienceId}": {
            "get": {
                "description": "Returns audience by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audience"
                ],
                "summary": "Get audience",
                "parameters": [
                    {
                        "type": "string",
                        "description": "audienceId",
                        "name": "audienceId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AudienceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates an audience",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audience"
                ],
                "summary": "Update audience",
                "parameters": [
                    {
                        "type": "string",
                        "description": "audienceId",
                        "name": "audienceId",
                        "in": "path"
                    },
                    {
                        "description": "audienceId",
                        "name": "command",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/response.AudienceCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes an audience",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audience"
                ],
                "summary": "Delete audience",
                "parameters": [
                    {
                        "type": "string",
                        "description": "audienceId",
                        "name": "audienceId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/audience/{audienceId}/favorite": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Favorites an audience",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audience"
                ],
                "summary": "Favorite audience",
                "parameters": [
                    {
                        "type": "string",
                        "description": "audienceId",
                        "name": "audienceId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Unfavorites an audience",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audience"
                ],
                "summary": "Unfavorite audience",
                "parameters": [
                    {
                        "type": "string",
                        "description": "audienceId",
                        "name": "audienceId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.AudienceCommand": {
            "type": "object",
            "properties": {
                "ageGroupMax": {
                    "type": "integer"
                },
                "ageGroupMin": {
                    "type": "integer"
                },
                "birthCountry": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "statType": {
                    "$ref": "#/definitions/response.AudienceStatTypeCommand"
                },
                "statTypeValue": {
                    "type": "number"
                }
            }
        },
        "response.AudiencePageResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.AudienceResponse"
                    }
                }
            }
        },
        "response.AudienceResponse": {
            "type": "object",
            "properties": {
                "ageGroupMax": {
                    "type": "integer"
                },
                "ageGroupMin": {
                    "type": "integer"
                },
                "assetId": {
                    "type": "string"
                },
                "birthCountry": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "descriptionFormatted": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "statType": {
                    "$ref": "#/definitions/response.AudienceStatTypeResponse"
                },
                "statTypeValue": {
                    "type": "number"
                }
            }
        },
        "response.AudienceStatTypeCommand": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "titleFormatted": {
                    "type": "string"
                }
            }
        },
        "response.AudienceStatTypeResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "titleFormatted": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}